move config tests to their own file
