+++ jupyter-extension/jupyterlab_pachyderm/tests/test_integrations.py
@@ -17,7 +17,7 @@ from pachyderm_sdk import Client
 from pachyderm_sdk.api import pfs, pps
 from pachyderm_sdk.config import ConfigFile
 
-from jupyterlab_pachyderm.tests import TEST_NOTEBOOK
+from jupyterlab_pachyderm.tests import DEFAULT_PROJECT
 
 DEFAULT_PROJECT = "default"
 
@@ -486,66 +486,3 @@ async def test_download_datum(pachyderm_resources, http_client: AsyncClient):
     ) == sorted(files)
     assert f"{repos[1]}_dev" in list(os.walk(PFS_MOUNT_DIR))[0][1]
     assert len(list(os.walk(os.path.join(PFS_MOUNT_DIR, repos[2])))[0][2]) == 1
-
-
-class TestConfigHandler:
-
-    @staticmethod
-    @pytest.mark.no_config
-    async def test_config_no_file(app: Application, http_client: AsyncClient):
-        """Test that if there is no config file present, the extension does not
-        use a default config."""
-        # Arrange
-        config_file = app.settings.get("pachyderm_config_file")
-        assert config_file is not None and not config_file.exists()
-
-        # Act
-        response = await http_client.get("/config")
-
-        # Assert
-        response.raise_for_status()
-        payload = response.json()
-        assert payload["cluster_status"] == "INVALID"
-        assert payload["pachd_address"] == ""
-
-    @staticmethod
-    @pytest.mark.no_config
-    async def test_do_not_set_invalid_config(app: Application, http_client: AsyncClient):
-        """Test that PUT /config does not store an invalid config."""
-        # Arrange
-        assert app.settings.get("pachyderm_client") is None
-        invalid_address = "localhost:33333"
-        data = {"pachd_address": invalid_address}
-
-        # Act
-        response = await http_client.put("/config", json=data)
-
-        # Assert
-        response.raise_for_status()
-        payload = response.json()
-        assert payload["cluster_status"] == "INVALID"
-        assert payload["pachd_address"] == invalid_address
-        assert app.settings.get("pachyderm_client") is None
-
-    @staticmethod
-    async def test_config(pach_config: Path, http_client: AsyncClient):
-        """Test that PUT /config with a valid configuration is processed
-        by the ConfigHandler as expected."""
-        # PUT request
-        local_active_context = ConfigFile.from_path(PACH_CONFIG).active_context
-
-        payload = {"pachd_address": local_active_context.pachd_address}
-        r = await http_client.put("/config", json=payload)
-
-        config = ConfigFile.from_path(pach_config)
-        new_active_context = config.active_context
-
-        assert r.status_code == 200, r.text
-        assert r.json()["cluster_status"] != "INVALID"
-        assert r.json()["pachd_address"] == new_active_context.pachd_address
-
-        # GET request
-        r = await http_client.get("/config")
-
-        assert r.status_code == 200, r.text
-        assert r.json()["cluster_status"] != "INVALID"
